// src/hooks/models/interfaces/uncompletedParkingProcess.interface.ts
export interface IUncompletedParkingProcess {
  _id: string;
  parking: {
    _id: string;
    title: string;
  };
  driver: {
    _id: string;
    carPlate: string;
  };
  payment: {
    currency: string;
    value: null;
  };
  entryCarTime: string;
  departureCarTime: null;
  isCompleted: false;
}